Vorpal Bunnies! / ThoughtDefs / Hediffs
Added the Vorpal Bunny and several variants with differing abilities and textures.

Added 3 new thought defs to go along with a few of the vorpal bunnies.

Gave each Gnomaclu variant a 50% chance to explode on damage instead of the old 1-shot-kill every time.

FINALLY, for real this time, fixed the damn fire stream textures to be actual png's with no background for the blazing felidae verb.

Added 7 new hediffs for stages that were meant to be used for the vorpal bunnies but didn't work out as planned. Maybe they can be used for a future animal though. :P